+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-All notable changes are documented in this file.
+Notable changes to the CLI and official extensions are documented in this file.
+
 The sections should follow the order `Packaging`, `Added`, `Changed`, `Fixed` and `Removed`.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).