+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [5.1.0] - 2023-05-04
+
 ### Added
 - CLI will look for the corresponding lockfile when analyzing a manifest file
 - Allow analyzing manifest files by generating lockfiles on-demand
@@ -490,7 +492,8 @@ before, the existing project ID will be re-linked.
 ## 0.0.1
 - Initial release.
 
-[unreleased]: https://github.com/phylum-dev/cli/compare/v5.0.1...HEAD
+[unreleased]: https://github.com/phylum-dev/cli/compare/v5.1.0...HEAD
+[5.1.0]: https://github.com/phylum-dev/cli/compare/v5.0.1...v5.1.0
 [5.0.1]: https://github.com/phylum-dev/cli/compare/v5.0.0...v5.0.1
 [5.0.0]: https://github.com/phylum-dev/cli/compare/v4.8.0...v5.0.0
 [4.8.0]: https://github.com/phylum-dev/cli/compare/v4.7.0...v4.8.0